features added

+++ src/main/java/tk/kaicheng/controllers/ProfileController.java
@@ -2,13 +2,19 @@ package tk.kaicheng.controllers;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import tk.kaicheng.models.Profile;
+import tk.kaicheng.models.User;
 import tk.kaicheng.services.ProfileService;
+import tk.kaicheng.services.UserService;
+
+import java.util.List;
 
 
 /**
@@ -18,27 +24,51 @@ import tk.kaicheng.services.ProfileService;
 public class ProfileController {
     @Autowired
     private ProfileService profileService;
+    @Autowired
+    private UserService userService;
 
+    // get user from security context
+
+    private User getContextUser(){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.findByUserName(auth.getName());
+        return user;
+    }
     @RequestMapping(value = "/user/profiles", method = RequestMethod.GET)
     public String list(Model model){
-        model.addAttribute("profiles", profileService.findAll());
+        User user = getContextUser();
+        // get profiles of the user
+        List<Profile> profiles = profileService.findByUser(user);
+        model.addAttribute("profiles", profiles);
         return "/user/profiles";
     }
 
     @RequestMapping(value = "/user/profile/{id}", method = RequestMethod.GET)
     public String showProfile(@PathVariable Integer id, Model model){
-        model.addAttribute("profile", profileService.findOne(id));
+        User user = getContextUser();
+        // show profile belonging to the user
+        Profile profile = profileService.findOne(id);
+        if(profile.getUser() != user) return "redirect:/403";
+        model.addAttribute("profile", profile);
         return "/user/profileshow";
     }
 
     @RequestMapping(value = "/user/profile/edit/{id}", method = RequestMethod.GET)
     public String edit(@PathVariable Integer id, Model model){
-        model.addAttribute("profile", profileService.findOne(id));
+        User user = getContextUser();
+        // edit profile belonging to the user
+        Profile profile = profileService.findOne(id);
+        if(profile.getUser() != user) return "redirect:/403";
+        model.addAttribute("profile", profile);
         return "/user/profileform";
     }
 
     @RequestMapping(value = "/user/profile/delete/{id}", method = RequestMethod.GET)
     public String delete(@PathVariable Integer id, Model model){
+        User user = getContextUser();
+        // delete profile belonging to the user
+        Profile profile = profileService.findOne(id);
+        if(profile.getUser() != user) return "redirect:/403";
         profileService.delete(id);
         return "redirect:/user/profiles";
     }
@@ -51,8 +81,12 @@ public class ProfileController {
 
     @RequestMapping(value = "/user/profile", method = RequestMethod.POST)
     public String saveProfile(Profile profile) {
-        if(profileService.findOne(profile.getId()) == null)
+        User user = getContextUser();
+        profile.setUser(user);
+        Profile fetchedProfile = profileService.findOne(profile.getId());
+        if( fetchedProfile == null) {
             profileService.save(profile);
+        }
         else
             profileService.updateProfileName(profile.getId(), profile.getProfileName());
         return "redirect:/user/profile/" + profile.getId();
